Very rudimentary flush manager
